+++ NEWS.md
@@ -2,7 +2,7 @@
 
 CHANGES
 
-* `makepredictcall(<dw_transformer>)` now supports `winsorize()`.
+* `makepredictcall(<dw_transformer>)` now supports `winsorize()`, `reverse()`, and `slide()`.
 
 * `data_read()` now also reads zip-files from URLs (#682).